+++ Build-LanguageServer.ps1
@@ -1,8 +1,16 @@
-﻿$dotnet = Get-Command 'dotnet'
+﻿Param(
+    [Parameter(Mandatory=$true)]
+    [string] $VersionPrefix = '0.2.0',
+
+    [Parameter()]
+    [string] $VersionSuffix = ''
+)
+
+$dotnet = Get-Command 'dotnet'
 
 $serverRoot = Join-Path $PSScriptRoot 'lib\server'
 $publishRoot = Join-Path $PSScriptRoot 'out'
 
-& $dotnet restore "$serverRoot\MSBuildProjectTools.sln"
-& $dotnet publish "$serverRoot\src\LanguageServer\LanguageServer.csproj" -f netcoreapp2.0 -o "$publishRoot\language-server"
-& $dotnet publish "$serverRoot\src\LanguageServer.TaskReflection\LanguageServer.TaskReflection.csproj" -f netcoreapp2.0 -o "$publishRoot\task-reflection"
+& $dotnet restore "$serverRoot\MSBuildProjectTools.sln" /p:VersionPrefix="$VersionPrefix" /p:VersionSuffix="$($VersionSuffix)"
+& $dotnet publish "$serverRoot\src\LanguageServer\LanguageServer.csproj" -f netcoreapp2.0 -o "$publishRoot\language-server" /p:VersionPrefix="$VersionPrefix" /p:VersionSuffix="$($VersionSuffix)"
+& $dotnet publish "$serverRoot\src\LanguageServer.TaskReflection\LanguageServer.TaskReflection.csproj" -f netcoreapp2.0 -o "$publishRoot\task-reflection" /p:VersionPrefix="$VersionPrefix" /p:VersionSuffix="$($VersionSuffix)"